// packages/datadog-api-client-v2/models/User.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserAttributes } from './UserAttributes';
import { UserResponseRelationships } from './UserResponseRelationships';
import { UsersType } from './UsersType';
/**
* User object returned by the API.
*/
export declare class User {
    'attributes'?: UserAttributes;
    /**
    * ID of the user.
    */
    'id'?: string;
    'relationships'?: UserResponseRelationships;
    'type'?: UsersType;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): User;
    static serialize(data: User): {
        [key: string]: any;
    };
    constructor();
}